+++ src/builder/bld.rs
@@ -3,6 +3,8 @@ use std::io::fs::{mkdir_recursive};
 
 use config::builders::Builder;
 use config::builders::Builder as B;
+use container::util::{clean_dir, copy_dir};
+use container::vagga::container_ver;
 
 use super::context::BuildContext;
 use super::commands::debian;
@@ -11,16 +13,90 @@ use super::commands::generic;
 use super::commands::pip;
 use super::commands::npm;
 use super::tarcmd;
-use container::util::clean_dir;
 use super::context::Distribution as Distr;
 
 
 pub trait BuildCommand {
+    fn configure(&self, ctx: &mut BuildContext) -> Result<(), String>;
     fn build(&self, ctx: &mut BuildContext) -> Result<(), String>;
 }
 
 
 impl BuildCommand for Builder {
+    fn configure(&self, ctx: &mut BuildContext) -> Result<(), String> {
+        match self {
+            &B::Container(ref name) => {
+                let cont = ctx.config.containers.get(name)
+                    .expect("Subcontainer not found");  // TODO
+                for b in cont.setup.iter() {
+                    try!(b.configure(ctx));
+                }
+            }
+            &B::Ubuntu(ref name) => {
+                if let Distr::Unknown = ctx.distribution {
+                    ctx.distribution = Distr::Ubuntu(debian::UbuntuInfo {
+                        release: name.to_string(),
+                        apt_update: true,
+                        has_universe: false,
+                    });
+                } else {
+                    return Err(format!("Conflicting distribution"));
+                };
+                try!(ctx.add_cache_dir(Path::new("/var/cache/apt"),
+                                       "apt-cache".to_string()));
+                try!(ctx.add_cache_dir(Path::new("/var/lib/apt/lists"),
+                                      "apt-lists".to_string()));
+                ctx.environ.insert("DEBIAN_FRONTEND".to_string(),
+                                   "noninteractive".to_string());
+                ctx.environ.insert("LANG".to_string(),
+                                   "en_US.UTF-8".to_string());
+            }
+            &B::UbuntuUniverse => {
+                match ctx.distribution {
+                    Distr::Ubuntu(ref mut ubuntu) => {
+                        ubuntu.has_universe = true;
+                    }
+                    _ => unreachable!(),
+                }
+            }
+            &B::Env(ref pairs) => {
+                for (k, v) in pairs.iter() {
+                    ctx.environ.insert(k.clone(), v.clone());
+                }
+            }
+            &B::CacheDirs(ref pairs) => {
+                for (k, v) in pairs.iter() {
+                    try!(ctx.add_cache_dir(k.clone(), v.clone()));
+                }
+            }
+            &B::Alpine(ref version) => {
+                if let Distr::Unknown = ctx.distribution {
+                    ctx.distribution = Distr::Alpine(alpine::AlpineInfo {
+                        mirror: alpine::choose_mirror(),
+                        version: version.to_string(),
+                    });
+                } else {
+                    return Err(format!("Conflicting distribution"));
+                };
+                try!(ctx.add_cache_dir(Path::new("/etc/apk/cache"),
+                                       "alpine-cache".to_string()));
+            }
+            &B::PipConfig(ref pip_settings) => {
+                ctx.pip_settings = pip_settings.clone();
+            }
+            &B::Py2Install(_) |
+            &B::Py3Install(_) |
+            &B::Py2Requirements(_) |
+            &B::Py3Requirements(_) => {
+                try!(ctx.add_cache_dir(Path::new("/tmp/pip-cache"),
+                                       "pip-cache".to_string()));
+                ctx.environ.insert("PIP_DOWNLOAD_CACHE".to_string(),
+                                   "/tmp/pip-cache".to_string());
+            }
+            _ => {}
+        }
+        Ok(())
+    }
     fn build(&self, ctx: &mut BuildContext) -> Result<(), String> {
         match self {
 
@@ -47,6 +123,13 @@ impl BuildCommand for Builder {
                     Distr::Alpine(_) => alpine::install(ctx, pkgs),
                 }
             }
+            &B::Container(ref name) => {
+                let version = try!(container_ver(name));
+                let path = Path::new("/vagga/base/.roots")
+                    .join(version).join("root");
+                try!(copy_dir(&path, &Path::new("/vagga/root")));
+                Ok(())
+            }
 
 
             &B::Ubuntu(ref name) => {
@@ -68,9 +151,6 @@ impl BuildCommand for Builder {
                 generic::run_command(ctx, cmd.as_slice())
             }
             &B::Env(ref pairs) => {
-                for (k, v) in pairs.iter() {
-                    ctx.environ.insert(k.clone(), v.clone());
-                }
                 Ok(())
             }
             &B::Remove(ref path) => {
@@ -92,9 +172,7 @@ impl BuildCommand for Builder {
                 Ok(())
             }
             &B::CacheDirs(ref pairs) => {
-                for (k, v) in pairs.iter() {
-                    try!(ctx.add_cache_dir(k.clone(), v.clone()));
-                }
+                // Done in configure
                 Ok(())
             }
             &B::Depends(_) => {
@@ -110,7 +188,6 @@ impl BuildCommand for Builder {
                 alpine::setup_base(ctx, name)
             }
             &B::PipConfig(ref pip_settings) => {
-                ctx.pip_settings = pip_settings.clone();
                 Ok(())
             }
             &B::Py2Install(ref pkgs) => {